+++ ak_library_management/__manifest__.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
 {
     "name": "Library",
-    "version": "18.0.1.6.0",
+    "version": "18.0.1.8.0",
     "license": "Other proprietary",
-    "depends": ["sale_management","stock","base_automation"],
+    "depends": ["sale_management","hr","stock","base_automation"],
     "author":"Raj",
     "website": "https://www.aktivsoftware.com",
     "data": [
-        # "security/security.xml",
+        "security/security.xml",
         "security/ir.model.access.csv",
+        "report/library_report_views.xml",
         "views/library_book_tags_views.xml",
         "views/library_book_category.xml",
         "views/library_library_views.xml",
@@ -20,11 +21,12 @@
         "views/res_partner_views.xml",
         "views/sale_order_views.xml",
         "views/borrow_transaction_history_views.xml",
-        # "views/stock_warehouse_views.xml",
+        "views/stock_warehouse_views.xml",
         "views/library_menu_views.xml",
         "data/ir_sequence_data.xml",
         "data/ir_cron_data.xml",
         "data/ir_action_data.xml",
+        "data/mail_template_data.xml",
         "wizard/sale_order_wizard_views.xml",
         "wizard/borrow_transaction_history_wizard_views.xml"
     ],