+++ ak_library_management/models/stock_warehouse.py
@@ -9,11 +9,11 @@ class StockWarehouse(models.Model):
     _inherit = 'stock.warehouse'
 
     library_assistant = fields.Many2one(
-        'hr.employee',
+        comodel_name='hr.employee',
         string="Library Assistant"
     )
     workers_ids = fields.Many2many(
-        'hr.employee',
+        comodel_name='hr.employee',
         string="Worker"
     )
 
@@ -22,7 +22,7 @@ class StockWarehouse(models.Model):
         if self.search_count([('library_assistant', '=', self.library_assistant.id)]) > 1:
             raise ValidationError("A Library Assistant can be assigned to only one warehouse.")
 
-    # @api.constrains('workers_ids')
-    # def _check_unique_workers(self):
-    #     if len(set(self.workers_ids.ids)) < len(self.workers_ids.ids):
-    #         raise ValidationError("Duplicate Workers are not allowed.")
+    @api.constrains('workers_ids')
+    def _check_unique_workers(self):
+        if any(rec.library_assistant in rec.workers_ids for rec in self if rec.library_assistant):
+            raise ValidationError("Assistant can not be Worker.")